requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==4.0.0
pymongo==4.3.3
selenium==4.10.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.0
openai==1.12.0
schedule==1.2.1
transformers==4.38.2
torch==2.6.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1 